created hash_token file
