+++ .env.example
@@ -4,5 +4,6 @@ SMTP_PASSWORD="example"
 SMTP_HOST="smtp.example.com"
 OTP_SECRET="lorememsum"
 JWT_SECRET="yoursecert"
+HMAC_SECRET="your-HMAC-secret-key"
 PASS_RESET_LINK="http://localhost:5173/reset-password"
 ALLOWED_ORIGINS="http://localhost:3000,http://yourdomain.com"
