+++ src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod constants;
+pub mod hash_token;
 pub mod pass_reset;
 pub mod scripts;
 pub mod verify_email;
